프론트엔드 파일시스템 모듈 리팩토링
