+++ src/lib/modules/filesystem/index.ts
@@ -1,0 +1,4 @@
+export * from "./category";
+export * from "./directory";
+export * from "./file";
+export * from "./types";
