+++ src/routes/(fullscreen)/settings/thumbnail/File.svelte
@@ -13,7 +13,7 @@
   import type { Writable } from "svelte/store";
   import { ActionEntryButton } from "$lib/components/atoms";
   import { DirectoryEntryLabel } from "$lib/components/molecules";
-  import type { FileInfo } from "$lib/modules/filesystem2.svelte";
+  import type { FileInfo } from "$lib/modules/filesystem";
   import { formatDateTime } from "$lib/utils";
   import type { GenerationStatus } from "./service.svelte";
